refactor: support/report 도메인 레포지토리 계층 분리
순수 도메인 인터페이스를 domain 패키지로 분리하여
프레임워크 독립성을 확보하고 의존성 역전 원칙을 준수합니다.

변경사항:
- UserReportRepository 순수 인터페이스 생성 (domain 패키지)
- ReviewReportRepository 순수 인터페이스 생성 (domain 패키지)
- JpaUserReportRepository로 이름 변경 및 @JpaRepositoryImpl 추가
- JpaReviewReportRepository로 이름 변경 및 @JpaRepositoryImpl 추가
- 서비스 계층의 import 경로를 domain 패키지로 업데이트
