+++ bottlenote-mono/src/main/java/app/bottlenote/support/report/repository/JpaReviewReportRepository.java
@@ -1,14 +1,16 @@
 package app.bottlenote.support.report.repository;
 
+import app.bottlenote.common.annotation.JpaRepositoryImpl;
 import app.bottlenote.support.report.domain.ReviewReport;
+import app.bottlenote.support.report.domain.ReviewReportRepository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public interface ReviewReportRepository extends JpaRepository<ReviewReport, Long> {
+@JpaRepositoryImpl
+public interface JpaReviewReportRepository
+    extends ReviewReportRepository, JpaRepository<ReviewReport, Long> {
 
   Optional<Object> findByUserIdAndReviewId(Long currentUserId, Long reportReviewId);
