+++ bottlenote-product-api/src/test/java/app/bottlenote/support/report/integration/ReportIntegrationTest.java
@@ -21,11 +21,11 @@ import app.bottlenote.review.domain.Review;
 import app.bottlenote.review.domain.ReviewRepository;
 import app.bottlenote.support.report.constant.UserReportType;
 import app.bottlenote.support.report.domain.ReviewReport;
+import app.bottlenote.support.report.domain.ReviewReportRepository;
 import app.bottlenote.support.report.dto.request.ReviewReportRequest;
 import app.bottlenote.support.report.dto.request.UserReportRequest;
 import app.bottlenote.support.report.dto.response.ReviewReportResponse;
 import app.bottlenote.support.report.dto.response.UserReportResponse;
-import app.bottlenote.support.report.repository.ReviewReportRepository;
 import app.bottlenote.user.constant.SocialType;
 import app.bottlenote.user.dto.request.OauthRequest;
 import com.fasterxml.jackson.databind.JsonNode;
